+++ .github/ISSUE_TEMPLATE/istio_update.md
@@ -20,7 +20,7 @@ ACs:
 - [ ] Tests and documentation updated if needed.
 - [ ] Verify that sidecars are in sync with Control Plane.
 - [ ] Istio and Envoy Version updated in the [`README.md`](https://github.com/kyma-project/istio) and [`/docs/user/README.md`](https://github.com/kyma-project/istio) files. You can use the scripts `scripts/get_module_istio_version.sh` and `scripts/get_module_envoy_version.sh` to extract the versions.
-- [ ] Check **compatibilityVersion** of the previous minor version . You can find it in the [`helm-profiles`](https://github.com/istio/istio/tree/master/manifests/helm-profiles) directory.
+- [ ] Check **compatibilityVersion** of the previous minor version . You can find it in the [`helm-profiles`](https://github.com/istio/istio/tree/master/manifests/helm-profiles) directory. Evaluate content, and adjust compatibility mode implementation. You can find it in the v1alpha2 package(api/v1alpha2/).
 - [ ] Check compatibility variables not covered by **compatibilityVersion** parameter for the new version in the Istio release notes. Update them if necessary.
 
 **DoD:**
