+++ api/v1alpha2/merge_test.go
@@ -539,101 +539,6 @@ var _ = Describe("Merge", func() {
 		})
 	})
 
-	Context("Compatibility Mode", func() {
-		It("should set compatibility variables on Istio Pilot when compatibility mode is on", func() {
-			//given
-			iop := iopv1alpha1.IstioOperator{
-				Spec: &operatorv1alpha1.IstioOperatorSpec{},
-			}
-			istioCR := Istio{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{},
-				},
-				Spec: IstioSpec{
-					CompatibilityMode: true,
-				},
-			}
-
-			// when
-			out, err := istioCR.MergeInto(iop)
-
-			//then
-			Expect(err).ShouldNot(HaveOccurred())
-
-			existingEnvs := map[string]string{}
-			for _, v := range out.Spec.Components.Pilot.K8S.GetEnv() {
-				existingEnvs[v.Name] = v.Value
-			}
-
-			for k, v := range pilotCompatibilityEnvVars {
-				Expect(existingEnvs[k]).To(Equal(v))
-			}
-
-		})
-
-		It("should set compatibility variables on Istio Pilot when compatibility mode is on despite disable external name aliast annotation set to false", func() {
-			//given
-			iop := iopv1alpha1.IstioOperator{
-				Spec: &operatorv1alpha1.IstioOperatorSpec{},
-			}
-			istioCR := Istio{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						"istio-operator.kyma-project.io/disable-external-name-alias": "false",
-					},
-				},
-				Spec: IstioSpec{
-					CompatibilityMode: true,
-				},
-			}
-
-			// when
-			out, err := istioCR.MergeInto(iop)
-
-			//then
-			Expect(err).ShouldNot(HaveOccurred())
-
-			existingEnvs := map[string]string{}
-			for _, v := range out.Spec.Components.Pilot.K8S.GetEnv() {
-				existingEnvs[v.Name] = v.Value
-			}
-
-			for k, v := range pilotCompatibilityEnvVars {
-				Expect(existingEnvs[k]).To(Equal(v))
-			}
-		})
-
-		It("should not set compatibility variables on Istio Pilot when compatibility mode is off", func() {
-			//given
-			iop := iopv1alpha1.IstioOperator{
-				Spec: &operatorv1alpha1.IstioOperatorSpec{},
-			}
-			istioCR := Istio{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{},
-				},
-				Spec: IstioSpec{
-					CompatibilityMode: false,
-				},
-			}
-
-			// when
-			out, err := istioCR.MergeInto(iop)
-
-			//then
-			Expect(err).ShouldNot(HaveOccurred())
-
-			variableCounter := 0
-			for _, value := range out.Spec.Components.Pilot.K8S.GetEnv() {
-				if v, ok := pilotCompatibilityEnvVars[value.Name]; ok && value.Value == v {
-					variableCounter++
-				}
-			}
-
-			Expect(variableCounter).To(Equal(0))
-		})
-	})
-
 	Context("IngressGateway", func() {
 		Context("When Istio CR has 500m configured for CPU and 500Mi for memory limits", func() {
 			It("should set CPU limits to 500m and 500Mi for memory in IOP", func() {
